Jussihakosalo/indica311 (#328)
* feat: first forays into updating Indica

* feat: 3.11 compatibility

* feat: passed precommit

* fix: disabled mypy precommit due to system error

* readme: edited readme

* fix: ppts reader data ambiguity

* Fixed README.md for installation guidelines and consequent .lock file

* Fixed 3.11 DataArray.data in another few instances

* Deleted print statement

* Upgraded default object .pkl to new versions.

* Deleted print statement

* feat: updated poetry

* feat: version fix

* Marcosertoli/311 further dataarray harmonisation (#329)

* Now consistent with new DataArray convention + assigning Indica-native attributes also to DataArray coordinates.

* Modified all DataArray coordinate instances to new convention

* Fixed bug in PPTS reader (#330)

* fix: rebuilt poetry lock

* Adding example runs for diagnostic models (#331)

* Marcosertoli/fix equilibrium r shift (#333)

* Upgraded python version in poetry dev dependencies & ipython

* Previous implementation was returning rhos with NaNs as fill_value which was making inversion methods crash.

* Github CI workflows now asking for python 3.11

* Update utilities.py

BAsed on comment

* Update abstractreader.py

Removed extra prints

---------

+++ indica/converters/abstractconverter.py
@@ -490,7 +490,7 @@ class CoordinateTransform(ABC):
 
         if hasattr(self, "equilibrium"):
             if t is None:
-                t = np.float(np.mean(self.equilibrium.rho.t))
+                t = np.float64(np.mean(self.equilibrium.rho.t))
             equil_bounds, angles, rho_equil = self.get_equilibrium_boundaries(t)
             x_ax = self.equilibrium.rmag.sel(t=t, method="nearest").values * np.cos(
                 angles
